B2v2 (round 5): press-edge session intent + bounded termination join
Review 5 blocker 2:
- NEW PendingSessionIntent (Core): immutable intent allocated SYNCHRONOUSLY
  at the press/toggle edge with generation + press timestamp + cancellation
  flag. Release/cancel invalidate it immediately, even before the queued
  begin operation starts (closing the window where a quick release queued
  behind a press could not preempt it).
- Controller: hotkey press, release, cancelSession and toggleManualSession
  all use the intent; beginSession(intent:) checks it before/after every
  await (history wait, model readiness, final pre-creation check) and aborts
  without preparing or activating the microphone when invalidated.
- DictationSession: NEW awaitTerminalAndReleased(deadlineNanosAhead:) — a
  bounded join (task-group race) on exactly-once terminal release; the
  release signal fires in finishTerminal. Controller termination now joins
  the session's release + run task before marking .sessionFinished (bounded;
  failure is logged, not silently claimed).
- Regression tests: intent validity/cancellation/generation; normal session
  reaches terminal release; live session not released within a short
  deadline; cancelled session releases.

+++ Sources/ZephyrFlowCore/AudioDrain.swift
@@ -198,12 +198,12 @@ public enum AudioDrainAssessment {
         public let barrierTimedOut: Bool
         public let barrierDrained: Bool
         public let consumerCompleted: Bool
-        public let lateAppends: Int
+        public let lateAppends: UInt64
         public let reconciled: Bool
         public init(
             seqDegraded: Bool, channelDegraded: Bool,
             barrierTimedOut: Bool, barrierDrained: Bool,
-            consumerCompleted: Bool, lateAppends: Int, reconciled: Bool
+            consumerCompleted: Bool, lateAppends: UInt64, reconciled: Bool
         ) {
             self.seqDegraded = seqDegraded
             self.channelDegraded = channelDegraded
